requirements.txt
albumentations>=1.3.1
opencv-python>=4.8.1.78
torch>=2.1.1+cu121
numpy>=1.24.3
pandas>=2.0.3
glog>=0.3.1
torchvision>=0.16.1+cu121
pillow>=10.2.0
dominate>=2.9.0
pyyaml>=6.0
yacs>=0.1.8
timm>=0.9.12
einops>=0.7.0
